added navbar component to test some nested view
