+++ src/App/components/shared/Navbar/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import wBooksLogo from '../../../../assets/wbooks_logo.svg';
+
+const Navbar = () =>
+  <nav>
+    <img src={wBooksLogo} />
+  </nav>;
+
+export default Navbar;
