+++ src/index.js
@@ -10,7 +10,7 @@ import './index.css';
 ReactDOM.render(
   <Provider store={store}>
     <App />
-  </Provider>
-  , document.getElementById('root')
+  </Provider>,
+  document.getElementById('root')
 );
 registerServiceWorker();
